Omit occluded windows
Omit any completely occluded windows from the list of candidate windows.
Partially-occluded windows are still considered; though in the future I
would prefer that only the visible portion of the window be considered.

Fixed a bug in which the last window returned by xdotool was excluded
from the list of candidate windows.